PPT Note Shift/pyproject.toml
[project]
dependencies = ["python-pptx~=1.0"]
